Remove Weightable and replace the Array implementation by IdentityToDouble
